+++ src/main/java/com/excilys/computerdatabase/computerdb/database/ICompanyDAO.java
@@ -16,9 +16,9 @@ public interface ICompanyDAO {
 
     Optional<Company> getCompanyById(long id) throws DaoException;
 
-    List<Pageable> getCompanyByName(String name, int limitStart, int size) throws DaoException;
+    List<Pageable> getCompanyByName(String name, long limitStart, long size) throws DaoException;
 
-    List<Pageable> getCompanys(int limitStart, int size) throws DaoException;
+    List<Pageable> getCompanys(long limitStart, long size) throws DaoException;
 
     long getNumberOfCompany() throws DaoException;
